// app/rich-text-editor/components/menus/types.ts
import type { Editor as CoreEditor } from '@tiptap/core'
import type { EditorState } from '@tiptap/pm/state'
import type { EditorView } from '@tiptap/pm/view'
import type { Editor } from '@tiptap/react'
import type React from 'react'

type MenuProps = {
  editor: Editor
  appendTo?: React.RefObject<any>
  shouldHide?: boolean
}

type ShouldShowProps = {
  editor?: CoreEditor
  view: EditorView
  state?: EditorState
  oldState?: EditorState
  from?: number
  to?: number
}

export type { MenuProps, ShouldShowProps }